+++ components/admin/Previews/PlaybookPreview/PlaybookPreview.js
@@ -1,0 +1,42 @@
+import PropTypes from 'prop-types';
+import { useQuery } from '@apollo/client';
+import { Loader } from 'semantic-ui-react';
+import ApolloError from 'components/errors/ApolloError';
+import ErrorSection from 'components/errors/ErrorSection';
+import Playbook from 'components/Playbook/Playbook';
+import { PLAYBOOK_QUERY } from 'lib/graphql/queries/playbook';
+import styles from './PlaybookPreview.module.scss';
+
+const PlaybookPreview = ( { id, item: itemFromServer } ) => {
+  const { data, error, loading } = useQuery( PLAYBOOK_QUERY, {
+    variables: { id },
+    skip: !!itemFromServer,
+  } );
+
+  if ( loading ) {
+    return (
+      <div className={ styles.loading }>
+        <Loader
+          active
+          inline="centered"
+          style={ { marginBottom: '1em' } }
+          content="Loading Playbook preview..."
+        />
+      </div>
+    );
+  }
+
+  if ( error ) return <ApolloError error={ error } />;
+  if ( !itemFromServer && !data?.playbook ) {
+    return <ErrorSection statusCode={ 404 } title="Preview is unavailable." />;
+  }
+
+  return <Playbook item={ itemFromServer || data.playbook } />;
+};
+
+PlaybookPreview.propTypes = {
+  id: PropTypes.string,
+  item: PropTypes.object,
+};
+
+export default PlaybookPreview;